Fix function signature

+++ attr.h
@@ -45,7 +45,7 @@ struct attr {
 int attr_init(struct attr *attr);
 void attr_free(struct attr *attr);
 ck_rv_t attr_add(struct attr *attr, ck_attribute_type_t type,
-                 void *value, unsigned long value_len, int dup);
+                 void *value, size_t value_len, int dup);
 int attr_match_template(struct attr *attr,
         struct ck_attribute *templ, unsigned long count);
 int attr_fill_template(struct attr *attr,
